requirements.txt
streamlit
sentence-transformers
scikit-learn
numpy
pandas
PyPDF2
python-docx
torch
